% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{combine}
\alias{combine}
\title{Combine time series}
\usage{
combine(ts1, ts2)
}
\arguments{
\item{ts1}{a time series}

\item{ts2}{a time series or numeric vector}
}
\value{
The combination of \code{ts1} and \code{ts2}
}
\description{
Combine time series
}
