% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{knn_examples}
\alias{knn_examples}
\title{Examples of the model associated with a prediction}
\usage{
knn_examples(forecast)
}
\arguments{
\item{forecast}{A \code{knnForecast} object.}
}
\value{
A matrix including the features and targets of the examples
   associated with the model of a \code{knnForecast} object.
}
\description{
It allows to see the examples of the model associated to a
\code{knnForecast} object.
}
\examples{
pred <- knn_forecasting(ts(1:8), h = 1, lags = 1:2, k = 2)
knn_examples(pred)
}
