% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{nearest_neighbors}
\alias{nearest_neighbors}
\title{Nearest neighbors associated with predictions}
\usage{
nearest_neighbors(forecast)
}
\arguments{
\item{forecast}{A \code{knnForecast} object.}
}
\value{
A list including the new instances used in KNN regression and their
nearest neighbors.
}
\description{
It allows to check the new instances and their nearest neighbors used in a
prediction associated with a "knnForecast" object.
}
\examples{
pred <- knn_forecasting(UKgas, h = 4, lags = 1:4, k = 2, msas = "MIMO")
nearest_neighbors(pred)
}
