% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_origin.R
\name{plot.knnForecastRO}
\alias{plot.knnForecastRO}
\title{Plot a prediction of a test set}
\usage{
\method{plot}{knnForecastRO}(x, h = NULL, ...)
}
\arguments{
\item{x}{the object obtained from a call to x \code{\link{rolling_origin}}.}

\item{h}{an integer. The forecasting horizon. If \code{NULL}, the maximum
forecasting horizon of all the test sets is used.}

\item{...}{Other plotting parameters to affect the plot.}
}
\description{
It uses a test set generated with the function \code{\link{rolling_origin}}
and plots its forecast.
}
