% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsfknn-package.R
\docType{package}
\name{tsfknn-package}
\alias{tsfknn}
\alias{tsfknn-package}
\title{tsfknn: Time Series Forecasting Using Nearest Neighbors}
\description{
Allows forecasting time series using nearest neighbors regression Francisco Martinez, Maria P. Frias, Maria D. Perez-Godoy and Antonio J. Rivera (2019) \doi{10.1007/s10462-017-9593-z}. When the forecasting horizon is higher than 1, two multi-step ahead forecasting strategies can be used. The model built is autoregressive, that is, it is only based on the observations of the time series. The nearest neighbors used in a prediction can be consulted and plotted.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/franciscomartinezdelrio/tsfknn}
  \item Report bugs at \url{https://github.com/franciscomartinezdelrio/tsfknn/issues}
}

}
\author{
\strong{Maintainer}: Francisco Martinez \email{fmartin@ujaen.es}

}
\keyword{internal}
