% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsfknn.R
\docType{package}
\name{tsfknn}
\alias{tsfknn}
\alias{tsfknn-package}
\title{tsfknn: A package for time series forecasting using KNN regression.}
\description{
The tsfknn package allows univariate time series forecasting using KNN
regression.
}
\section{Functions}{

\describe{
   \item{knnForecasting}{It is used to forecast a time series}
   \item{nearest_neighbors}{To see the nearest neighbors used to
          forecast a times series}
   \item{knn_examples}{To see the examples used by the KNN model}
   \item{autoplot}{To plot a prediction and the nearest neighbors
         used in the prediction}
}
}

